+++ components/query/query-view/src/main/java/uk/gov/justice/services/components/query/view/interceptors/QueryViewInterceptorChainProvider.java
@@ -3,15 +3,13 @@ package uk.gov.justice.services.components.query.view.interceptors;
 import static uk.gov.justice.services.core.annotation.Component.QUERY_VIEW;
 
 import uk.gov.justice.services.components.common.BaseInterceptorChainProvider;
-import uk.gov.justice.services.core.accesscontrol.LocalAccessControlInterceptor;
 import uk.gov.justice.services.core.audit.LocalAuditInterceptor;
-
-import org.apache.commons.lang3.tuple.ImmutablePair;
+import uk.gov.justice.services.core.interceptor.InterceptorChainEntry;
 
 public class QueryViewInterceptorChainProvider extends BaseInterceptorChainProvider {
 
     public QueryViewInterceptorChainProvider() {
-        interceptorChainTypes().add(new ImmutablePair<>(3000, LocalAuditInterceptor.class));
+        interceptorChainTypes().add(new InterceptorChainEntry(3000, LocalAuditInterceptor.class));
     }
 
     @Override