+++ framework-api/framework-api-core/src/main/java/uk/gov/justice/services/core/interceptor/InterceptorChainProvider.java
@@ -2,8 +2,6 @@ package uk.gov.justice.services.core.interceptor;
 
 import java.util.List;
 
-import org.apache.commons.lang3.tuple.Pair;
-
 
 public interface InterceptorChainProvider {
 
@@ -15,10 +13,10 @@ public interface InterceptorChainProvider {
     String component();
 
     /**
-     * Provide a List containing Pairs of Priority and Interceptor Class for the InterceptorCache to
+     * Provide a List containing {@link InterceptorChainEntry} for the InterceptorCache to
      * create an InterceptorChain.  Priority order is low is highest. e.g. 1 = is highest priority
      *
      * @return Deque containing Interceptor Classes
      */
-    List<Pair<Integer, Class<? extends Interceptor>>> interceptorChainTypes();
+    List<InterceptorChainEntry> interceptorChainTypes();
 }